+++ src/main/kotlin/com/github/saundefined/bitrix_idea/validation/UrlVerifier.kt
@@ -1,0 +1,7 @@
+package com.github.saundefined.bitrix_idea.validation
+
+class UrlVerifier {
+    fun verify(input: String): Boolean {
+        return input.matches(Regex("^https?://\\S*\$")) || input.isEmpty()
+    }
+}